Split binary_expr.rs into smaller modules
